Cleaned up GRIDAQThread code.
git-svn-id: file:///var/svn/grif/framework@221 d8f7f5d7-83b2-4a66-a7c0-045f5512b77b

+++ trunk/source/GRILoader.cpp
@@ -80,7 +80,7 @@ list<GRIParam*>* GRILoader::readNewParamList(QString filePath, list<GRIParam*>* 
 
                 //debug-------------------------------------------------------
 //                this->pauseProgram("reading tag 'PARAMETER'");
-                  QVariant size = paramList->size();
+                  QVariant size = (unsigned int) paramList->size();
                    paramList->push_back(this->readParameter(xml, size.toString()));
 //                    head->addChild(this->readParameter(xml, head));
             }
